Move geoblock config Prepare out of plugin.go so ServeHTTP stays the Plugin file.
