+++ knowledge/devdocs/core_geoblock_database_source.md
@@ -46,5 +46,5 @@ u, err := dbsource.Start(cfg, logger, onUpdate)
 ## Key files
 
 - `pkg/dbsource` — Config, Resolve, Update, unpack, Updater
-- `pkg/geoblock/plugin.go` — `databaseSources`, pointers, `catalogSource`
+- `pkg/geoblock/config.go` — `databaseSources`, pointers, `catalogSource`
 - `pkg/dbutils` — HTTP GET, `DatedKeyGlob`, `DownloadHint`
